refactor(pixelflow-pipeline): dedupe the JIT/oracle harness and journal writer; document the domain model
Two independent "one definition, imported, not restated" violations in
the measurement path, both named in
docs/plans/2026-08-17-cost-model-domain.md:

- J14: `run_scalar` (broadcast one coordinate to all lanes, read lane 0)
  existed as three byte-identical per-ISA copies in
  quarantine.rs, bench_extraction_3way.rs, and
  pixelflow-search/tests/prod_kernel_jit.rs — quarantine.rs's own comment
  admitted "Verbatim from ...". A fix landing in one copy and silently
  not propagating to the others is exactly the NO-SILENT-FAILURES class
  this repo forbids. Both binaries now import the one copy in the new
  `oracle_compare.rs`.
- J15: the research-journal append mechanics (serialize, mkdir the
  parent, refuse an unsmudged Git LFS pointer, append-or-panic) were two
  copies, one of them missing the LFS guard. The new `journal.rs` module
  is the one writer; `bench_extraction_3way` and `bootstrap_extraction_head`
  both call `journal::append_record` for their (structurally different,
  and staying different) record types.

Also adds the domain-model doc this round's reorganization follows: a
join of three independent analyses (domain description, name/argument
clustering across 1,330 fns, caller-verified live/dead disposition) into
15 candidate types plus 6 anti-rows the join says not to build, and the
ranked execution list this PR implements items 1/5/6/10 of (Guide
encapsulation, run_scalar dedup, saturation-loop dedup, journal dedup).
The remaining items — `Expr<'a>`, `Extraction` with its `EmbeddingsEpoch`
tag, `BenchRequest`, `Fence<T>`/`FenceKey`, `VariantSet`, `RoundVerdict`
— are denoted (§1 of the doc) but not yet built; §4 ranks what's next and
names what NOT to attempt ahead of its consumer.
