+++ pixelflow-pipeline/src/lib.rs
@@ -8,6 +8,8 @@
 //! to also host.
 
 pub mod jit_bench;
+pub mod journal;
+pub mod oracle_compare;
 pub mod oracle_lowering;
 
 // Training infrastructure (requires std feature)
